Pushed MetaData into it's own class
